% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch_app}
\alias{launch_app}
\title{launch_app
Run the graphical interface to the game in a web browser}
\usage{
launch_app()
}
\description{
launch_app
Run the graphical interface to the game in a web browser
}
