% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_mind.R
\name{score_df}
\alias{score_df}
\title{score_df}
\usage{
score_df(dt, true_word)
}
\arguments{
\item{dt}{The attempts data frame}

\item{true_word}{The word to be guessed}
}
\value{
A data frame of characters with scores (colors)
}
\description{
Calculates scores for the attempts data frame
}
