% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_mind.R
\name{score_row}
\alias{score_row}
\title{score_row}
\usage{
score_row(v_try, v_tru)
}
\arguments{
\item{v_try}{Vectorized version of the guess word}

\item{v_tru}{Vectorized version of the correct word}
}
\value{
A character vector of scores with the colors
}
\description{
Calculates score for a given guess word
}
